+++ src/Components/vitest.config.ts
@@ -49,6 +49,13 @@ export default mergeConfig(
             setupFiles: ['./.storybook/vitest.setup.ts'],
           },
         },
+        {
+          extends: true,
+          test: {
+            name: 'unit',
+            include: ['src/**/*.test.ts'],
+          },
+        },
       ],
     },
   }),