+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### ✨ Added
 
+-  Workspace component (by @Freymaurer)
 -  Add ARC-Selector that enables selection of recently used ARCs without searching in the folder structure
 -  Implement foundation for filetree and combine it with chokidar filewatcher
 -  Implement Javascript bindings for chockidar and simple git